split PassengerEngine into DefaultPassengerEngine and PassengerEngineWithPrebooking
