+++ contribs/dvrp/src/main/java/org/matsim/contrib/dvrp/passenger/PassengerHandler.java
@@ -1,0 +1,101 @@
+/*
+ * *********************************************************************** *
+ * project: org.matsim.*
+ * *********************************************************************** *
+ *                                                                         *
+ * copyright       : (C) 2020 by the members listed in the COPYING,        *
+ *                   LICENSE and WARRANTY file.                            *
+ * email           : info at matsim dot org                                *
+ *                                                                         *
+ * *********************************************************************** *
+ *                                                                         *
+ *   This program is free software; you can redistribute it and/or modify  *
+ *   it under the terms of the GNU General Public License as published by  *
+ *   the Free Software Foundation; either version 2 of the License, or     *
+ *   (at your option) any later version.                                   *
+ *   See also COPYING, LICENSE and WARRANTY file                           *
+ *                                                                         *
+ * *********************************************************************** *
+ */
+
+package org.matsim.contrib.dvrp.passenger;
+
+import static java.lang.String.format;
+
+import java.util.Set;
+import java.util.concurrent.atomic.AtomicInteger;
+
+import org.apache.log4j.Logger;
+import org.matsim.api.core.v01.Id;
+import org.matsim.api.core.v01.events.PersonEntersVehicleEvent;
+import org.matsim.api.core.v01.events.PersonLeavesVehicleEvent;
+import org.matsim.contrib.dvrp.optimizer.Request;
+import org.matsim.core.api.experimental.events.EventsManager;
+import org.matsim.core.mobsim.framework.MobsimDriverAgent;
+import org.matsim.core.mobsim.framework.MobsimPassengerAgent;
+import org.matsim.core.mobsim.qsim.InternalInterface;
+import org.matsim.core.mobsim.qsim.interfaces.MobsimVehicle;
+
+/**
+ * @author Michal Maciejewski (michalm)
+ */
+class PassengerHandler {
+	private static final Logger LOGGER = Logger.getLogger(PassengerHandler.class);
+
+	private final String mode;
+	private final EventsManager eventsManager;
+	private final AtomicInteger currentRequestId = new AtomicInteger(-1);
+
+	private InternalInterface internalInterface;
+
+	PassengerHandler(String mode, EventsManager eventsManager) {
+		this.mode = mode;
+		this.eventsManager = eventsManager;
+	}
+
+	void setInternalInterface(InternalInterface internalInterface) {
+		this.internalInterface = internalInterface;
+	}
+
+	Id<Request> createRequestId() {
+		return Id.create(mode + "_" + currentRequestId.incrementAndGet(), Request.class);
+	}
+
+	boolean validateRequest(PassengerRequest request, PassengerRequestValidator requestValidator, double now) {
+		Set<String> violations = requestValidator.validateRequest(request);
+		if (!violations.isEmpty()) {
+			String cause = String.join(", ", violations);
+			LOGGER.warn(format("Request: %s of mode: %s will not be served. Agent will get stuck. Cause: %s",
+					request.getId(), mode, cause));
+			eventsManager.processEvent(
+					new PassengerRequestRejectedEvent(now, mode, request.getId(), request.getPassengerId(), cause));
+		}
+		return violations.isEmpty();
+	}
+
+	boolean tryPickUpPassenger(MobsimDriverAgent driver, MobsimPassengerAgent passenger, double now) {
+		if (internalInterface.unregisterAdditionalAgentOnLink(passenger.getId(), driver.getCurrentLinkId()) == null) {
+			//only possible with prebooking
+			return false;
+		}
+
+		MobsimVehicle mobVehicle = driver.getVehicle();
+		mobVehicle.addPassenger(passenger);
+		passenger.setVehicle(mobVehicle);
+
+		eventsManager.processEvent(new PersonEntersVehicleEvent(now, passenger.getId(), mobVehicle.getId()));
+		return true;
+	}
+
+	void dropOffPassenger(MobsimDriverAgent driver, MobsimPassengerAgent passenger, double now) {
+		MobsimVehicle mobVehicle = driver.getVehicle();
+		mobVehicle.removePassenger(passenger);
+		passenger.setVehicle(null);
+
+		eventsManager.processEvent(new PersonLeavesVehicleEvent(now, passenger.getId(), mobVehicle.getId()));
+
+		passenger.notifyArrivalOnLinkByNonNetworkMode(passenger.getDestinationLinkId());
+		passenger.endLegAndComputeNextState(now);
+		internalInterface.arrangeNextAgentState(passenger);
+	}
+}
